% Auto-generated: do not edit by hand
\name{button}

\alias{button}

\title{Button component}

\description{

}

\usage{
button(children=NULL, id=NULL, className=NULL, text=NULL, type=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{className}{Character. }

\item{text}{Character. }

\item{type}{Logical | numeric | character | named list | unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

